+++ crates/rspack_plugin_javascript/src/plugin/impl_plugin_for_js_plugin.rs
