+++ crates/rspack_plugin_javascript/src/plugin/mod.rs
@@ -1,4 +1,4 @@
-pub mod impl_js_plugin;
+pub mod impl_plugin_for_js_plugin;
 pub mod infer_async_modules_plugin;
 
 use std::hash::Hash;
